function parse_truncate_table moved to the truncate.rs
